Refactor eth topology
