+++ boyar/topology/ethereum/topology_test.go
@@ -1,14 +1,13 @@
-package topology
+package ethereum
 
 import (
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/orbs-network/boyarin/boyar/topology/ethereum"
 	"github.com/orbs-network/boyarin/strelets"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
-func Test_ipToString(t *testing.T) {
+func Test_IpToString(t *testing.T) {
 	require.Equal(t, "172.10.93.43", IpToString([4]byte{172, 10, 93, 43}))
 }
 
@@ -17,11 +16,11 @@ const NODE_ADDESSS_2 = "d27e2e7398e2582f63d0800330010b3e58952ff6"
 const NODE_IP_1 = "172.10.93.43"
 const NODE_IP_2 = "172.10.93.92"
 
-func TestConvert(t *testing.T) {
+func Test_RawTopology_FederationNodes(t *testing.T) {
 	firstAddress, _ := common.NewMixedcaseAddressFromString(NODE_ADDRESS_1)
 	secondAddress, _ := common.NewMixedcaseAddressFromString(NODE_ADDESSS_2)
 
-	topology := Convert(&ethereum.RawTopology{
+	federationNodes := (&RawTopology{
 		IpAddresses: [][4]byte{
 			{172, 10, 93, 43},
 			{172, 10, 93, 92},
@@ -30,7 +29,7 @@ func TestConvert(t *testing.T) {
 			firstAddress.Address(),
 			secondAddress.Address(),
 		},
-	})
+	}).FederationNodes()
 
 	require.EqualValues(t, []*strelets.FederationNode{
 		{
@@ -41,5 +40,5 @@ func TestConvert(t *testing.T) {
 			IP:  NODE_IP_2,
 			Key: NODE_ADDESSS_2,
 		},
-	}, topology)
+	}, federationNodes)
 }
